refactor: login, loading page 관련 유틸 함수 처리 및 리팩토링
