+++ client/src/components/LoadingContainer/LoadingContainer.tsx
@@ -8,7 +8,7 @@ const LoadingContainer = () => {
   return (
     <S.Body>
       <img src={loadingCat} alt="로딩중" width="30%" />
-      <S.Text>Loading...</S.Text>
+      <S.Text>Loading</S.Text>
     </S.Body>
   );
 };
